major rewrite and some fundamental coverage improvements
Completely rewrites the egghatch core to apply some shellcode
heuristics. E.g., it is assumed that bytes after a "call imm32"
instruction is data, except when the call target is a pop instruction
where its operand is later used in a call statement (e.g., "call ebp").

This moves some basic block logic to the Block object and extends it to
support more (un)conditional jumps and calls. Additional logic to better
identify basic blocks is also added, namely, cross-references are now
split & merged into one coherent basic block list.

This commit also adds the "size" field to the "text" values due to which
it isn't backwards compatible with version 0.1, but that's fine.